HARMONY-817: Complete renaming to harmonyservices/service-example
