+++ .github/workflows/publish-image.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - main
+      - harmony-817-2
     tags:
       - v*
 env:
@@ -23,7 +24,9 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - name: Build image
-        run: docker build . --file Dockerfile --tag "${GITHUB_REPOSITORY}" --label "runnumber=${GITHUB_RUN_ID}"
+        run: |
+          mkdir -p deps
+          docker build . --file Dockerfile --tag "${GITHUB_REPOSITORY}" --label "runnumber=${GITHUB_RUN_ID}"
       - name: Log into registry
         run: echo "${{ secrets.DOCKERHUB }}" | docker login -u "${DOCKER_USER}" --password-stdin
       - name: Push image
