+++ bin/harmony_service_example
@@ -17,5 +17,5 @@ if [ -f .env ]; then
   args="$args --env-file=.env"
 fi
 
-docker run --rm -it -v $(pwd):/home $args harmony/gdal "$@"
+docker run --rm -it -v $(pwd):/home $args harmonyservices/service-example "$@"
